Make Corge build tree usable by clients
We can generate a CorgeConfig file in the build tree of Corge
so that clients do not need to install it before trying to build
a package against it. This is helpful for many development tasks
and workflows, but relies on setting include and other paths
appropriately.

Add generation of Config and Targets files for Corge build tree.
Recursion to subdirectories is moved before this generation, as
creation of the Targets file for the Build tree requires the exports
set to be present.

Rename CorgeConfig.cmake.in template variables for header and
lib paths so these can be set in main script without interfering
with the CMAKE_INSTALL_XXX variables directly. Set the new variables
as required for Build/Install trees before calling the
configure_package_config_file function. It is these variables
that cause the relative complexity of setup, but they can be
removed if full imported targets with usage requirements are
used.

Do not change way Quux dependency is found in CorgeConfig.
This means that clients trying to build against a build of
Corge need to have CMAKE_PREFIX_PATH/Quux_DIR set to point
to the appropriate location for Quux.

Since build trees are not relocatable anyway, the location of
Quux could be hardcoded into the refind (via find_package)
for the build tree if needed. In that case, it would be
easiest to have a separate "CorgeDependencies" file included
by CorgeConfig. This abstraction could make the setup of
the build (non-relocatable)/install(reloctable) versions
easier.